move reload to be inside the wrapper component
