+++ src/components/wrapper/wrapper.js
@@ -24,20 +24,6 @@ export const Wrapper = ({ displayGlobalToolbar, theme }) => {
     'experiment-tracking': '',
   });
 
-  // Reload state is to ensure it will call redirectLocation in FlowchartWrapper
-  // only when the page is reloaded.
-  const [reload, setReload] = useState(false);
-
-  useEffect(() => setReload(true), []);
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setReload(false);
-    }, 200);
-
-    return () => clearTimeout(timer);
-  }, []);
-
   const { data: versionData } = useApolloQuery(GET_VERSIONS, {
     client,
     skip: !displayGlobalToolbar,
@@ -78,7 +64,7 @@ export const Wrapper = ({ displayGlobalToolbar, theme }) => {
             )}
             <Switch>
               <Route exact path={sanitizedPathname}>
-                <FlowChartWrapper reload={reload} />
+                <FlowChartWrapper />
               </Route>
               <Route path={`${sanitizedPathname}experiment-tracking`}>
                 <ExperimentWrapper />
